add tv import tests, fix-up tv match tests with parsed year
