+++ src/nefarious/tests/importer/tv.py
@@ -1,0 +1,117 @@
+import os
+import logging
+
+from django.core.cache import cache
+from django.utils import timezone
+from requests import HTTPError
+
+from nefarious.models import WatchTVEpisode, WatchTVShow
+from nefarious.parsers.tv import TVParser
+from nefarious.quality import video_extensions
+from nefarious.tests.importer.base import ImporterBase
+
+
+class TVImporter(ImporterBase):
+    INGEST_DEPTH_MAX = 3
+
+    def ingest_path(self, file_path):
+        file_name = os.path.basename(file_path)
+
+        parser = TVParser(file_name)
+
+        # match
+        if parser.match:
+            file_extension_match = parser.file_extension_regex.search(file_name)
+            if file_extension_match:
+                title = parser.match['title']
+                # no title found so it could be sub-directories like:
+                #   show/season 01/s01e01.mkv
+                #   show - season 01/s01e01.mkv
+                if not title:
+                    if self._ingest_depth(file_path) > 0:
+
+                        # append the top most parent folder as the title, i.e "show/season 01/s01e01.mkv" would become "show - s01e01.mkv"
+                        file_path_split = file_path.split(os.sep)
+
+                        # possible parent directories
+                        parent_titles = [
+                            # "show - season 01/s01e01.mkv" would define title as "show - season 01"
+                            os.path.basename(os.sep.join(file_path_split[:-(self._ingest_depth(file_path))])),
+                            # "show/season 01/s01e01.mkv" would define title as "show - s01e01.mkv"
+                            '{} - {}'.format(
+                                os.path.basename(os.sep.join(file_path_split[:-(self._ingest_depth(file_path))])), file_name),
+                        ]
+                        for parent_title in parent_titles:
+                            parent_parser = TVParser(parent_title)
+                            # define the title and merge the parent and the file parser matches
+                            if parent_parser.match and parent_parser.match['title']:
+                                title = parent_parser.match['title']
+                                parser.match.update(parent_parser.match)
+                                break
+                        else:  # for/else
+                            logging.warning('[NO_MATCH_TITLE] Could not match nested file "{}"'.format(file_path))
+                            return False
+                    else:
+                        logging.warning('[NO_MATCH_TITLE] Could not match file without title "{}"'.format(file_path))
+                        return False
+                file_extension = file_extension_match.group()
+                if file_extension in video_extensions():
+                    if parser.is_single_episode():
+                        if WatchTVEpisode.objects.filter(download_path=file_path).exists():
+                            logging.info('[SKIP] skipping already-processed file "{}"'.format(file_path))
+                            return False
+                        # get or set tmdb search results for this title in the cache
+                        results = cache.get(title)
+                        if not results:
+                            try:
+                                results = self.tmdb_search.tv(query=title, language=self.nefarious_settings.language)
+                            except HTTPError:
+                                logging.error('[ERROR_TMDB] tmdb search exception for title {} on file "{}"'.format(title, file_path))
+                                return False
+                            cache.set(title, results, 60 * 60)
+                        # loop over results for the exact match
+                        for result in results['results']:
+                            poster_path = self.nefarious_settings.get_tmdb_poster_url(result['poster_path']) if result['poster_path'] else ''
+                            # normalize titles and see if they match
+                            if parser.normalize_media_title(result['name']) == title:
+                                season_number = parser.match['season'][0]
+                                episode_number = parser.match['episode'][0]
+                                watch_show, _ = WatchTVShow.objects.get_or_create(
+                                    tmdb_show_id=result['id'],
+                                    defaults=dict(
+                                        user=self.user,
+                                        name=result['name'],
+                                        poster_image_url=poster_path,
+                                    ),
+                                )
+                                episode_result = self.tmdb_client.TV_Episodes(result['id'], season_number, episode_number)
+                                try:
+                                    episode_data = episode_result.info()
+                                except HTTPError:
+                                    logging.error('[ERROR_TMDB] tmdb episode exception for title {} on file "{}"'.format(title, file_path))
+                                    return False
+                                watch_episode, _ = WatchTVEpisode.objects.update_or_create(
+                                    tmdb_episode_id=episode_data['id'],
+                                    defaults=dict(
+                                        user=self.user,
+                                        watch_tv_show=watch_show,
+                                        season_number=season_number,
+                                        episode_number=episode_number,
+                                        download_path=file_path,
+                                        collected=True,
+                                        collected_date=timezone.utc.localize(timezone.datetime.utcnow()),
+                                    ),
+                                )
+                                logging.info('[MATCH] Saved episode "{}" from file "{}"'.format(watch_episode, file_path))
+                                return watch_episode
+                        else:  # for/else
+                            logging.warning('[NO_MATCH_MEDIA] No media match for title "{}" and file "{}"'.format(title, file_path))
+                    else:
+                        logging.warning('[NO_MATCH_SINGLE_EPISODE] No single episode title match for title "{}" and file "{}"'.format(title, file_path))
+                else:
+                    logging.warning('[NO_MATCH_VIDEO] No valid video file extension for file "{}"'.format(file_path))
+            else:
+                logging.warning('[NO_MATCH_EXTENSION] No file extension for file "{}"'.format(file_path))
+        else:
+            logging.info('[NO_MATCH_UNKNOWN] Unknown match for file "{}"'.format(file_path))
+        return False
