+++ src/nefarious/tests/test_single_episode.py
@@ -15,21 +15,21 @@ class TVMatch(TestCase):
             ["Chuck.4x05.HDTV.XviD-LOL", "Chuck", 4, 5],
             ["The.Girls.Next.Door.S03E06.DVDRip.XviD-WiDE", "The Girls Next Door", 3, 6],
             ["Degrassi.S10E27.WS.DSR.XviD-2HD", "Degrassi", 10, 27],
-            ["Parenthood.2010.S02E14.HDTV.XviD-LOL", "Parenthood 2010", 2, 14],
+            ["Parenthood.2010.S02E14.HDTV.XviD-LOL", "Parenthood", 2, 14],
             ["Hawaii Five 0 S01E19 720p WEB DL DD5 1 H 264 NT", "Hawaii Five 0", 1, 19],
             ["The Event S01E14 A Message Back 720p WEB DL DD5 1 H264 SURFER", "The Event", 1, 14],
             ["Adam Hills In Gordon St Tonight S01E07 WS PDTV XviD FUtV", "Adam Hills In Gordon St Tonight", 1, 7],
             ["Adventure.Inc.S03E19.DVDRip.XviD-OSiTV", "Adventure Inc", 3, 19],
             ["S03E09 WS PDTV XviD FUtV", "", 3, 9],
             ["5x10 WS PDTV XviD FUtV", "", 5, 10],
-            ["Castle.2009.S01E14.HDTV.XviD-LOL", "Castle 2009", 1, 14],
-            ["Pride.and.Prejudice.1995.S03E20.HDTV.XviD-LOL", "Pride and Prejudice 1995", 3, 20],
+            ["Castle.2009.S01E14.HDTV.XviD-LOL", "Castle", 1, 14],
+            ["Pride.and.Prejudice.1995.S03E20.HDTV.XviD-LOL", "Pride and Prejudice", 3, 20],
             ["The.Office.S03E115.DVDRip.XviD-OSiTV", "The Office", 3, 115],
             ["Parks and Recreation - S02E21 - 94 Meetings - 720p TV.mkv", "Parks and Recreation", 2, 21],
             ["24-7 Penguins-Capitals- Road to the NHL Winter Classic - S01E03 - Episode 3.mkv", "24-7 Penguins-Capitals- Road to the NHL Winter Classic", 1, 3],
             ["Adventure.Inc.S03E19.DVDRip.\"XviD\"-OSiTV", "Adventure Inc", 3, 19],
-            ["Hawaii Five-0 (2010) - 1x05 - Nalowale (Forgotten/Missing)", "Hawaii Five-0 (2010)", 1, 5],
-            ["Hawaii Five-0 (2010) - 1x05 - Title", "Hawaii Five-0 (2010)", 1, 5],
+            ["Hawaii Five-0 (2010) - 1x05 - Nalowale (Forgotten/Missing)", "Hawaii Five-0", 1, 5],
+            ["Hawaii Five-0 (2010) - 1x05 - Title", "Hawaii Five-0", 1, 5],
             ["House - S06E13 - 5 to 9 [DVD]", "House", 6, 13],
             ["The Mentalist - S02E21 - 18-5-4", "The Mentalist", 2, 21],
             ["Breaking.In.S01E07.21.0.Jump.Street.720p.WEB-DL.DD5.1.h.264-KiNGS", "Breaking In", 1, 7],
@@ -44,8 +44,8 @@ class TVMatch(TestCase):
             ["The Mentalist S02E21 18 5 4 720p WEB DL DD5 1 h 264 EbP", "The Mentalist", 2, 21],
             ["01x04 - Halloween, Part 1 - 720p WEB-DL", "", 1, 4],
             ["extras.s03.e05.ws.dvdrip.xvid-m00tv", "extras", 3, 5],
-            ["castle.2009.416.hdtv-lol", "castle 2009", 4, 16],
-            ["hawaii.five-0.2010.217.hdtv-lol", "hawaii five-0 2010", 2, 17],
+            ["castle.2009.416.hdtv-lol", "castle", 4, 16],
+            ["hawaii.five-0.2010.217.hdtv-lol", "hawaii five-0", 2, 17],
             ["Looney Tunes - S1936E18 - I Love to Singa", "Looney Tunes", 1936, 18],
             ["American_Dad!_-_7x6_-_The_Scarlett_Getter_[SDTV]", "American Dad!", 7, 6],
             ["Falling_Skies_-_1x1_-_Live_and_Learn_[HDTV-720p]", "Falling Skies", 1, 1],
@@ -88,15 +88,15 @@ class TVMatch(TestCase):
             ["Constantine S1-E1-WEB-DL-1080p-NZBgeek", "Constantine", 1, 1],
             ["Constantine S1E1-WEB-DL-1080p-NZBgeek", "Constantine", 1, 1],
             ["NCIS.S010E16.720p.HDTV.X264-DIMENSION", "NCIS", 10, 16],
-            ["[ www.Torrenting.com ] - Revolution.2012.S02E17.720p.HDTV.X264-DIMENSION", "Revolution 2012", 2, 17],
-            ["Revolution.2012.S02E18.720p.HDTV.X264-DIMENSION.mkv", "Revolution 2012", 2, 18],
+            ["[ www.Torrenting.com ] - Revolution.2012.S02E17.720p.HDTV.X264-DIMENSION", "Revolution", 2, 17],
+            ["Revolution.2012.S02E18.720p.HDTV.X264-DIMENSION.mkv", "Revolution", 2, 18],
             ["Series - Season 1 - Episode 01 (Resolution).avi", "Series", 1, 1],
             ["5x09 - 100 [720p WEB-DL].mkv", "", 5, 9],
             ["1x03 - 274 [1080p BluRay].mkv", "", 1, 3],
             ["1x03 - The 112th Congress [1080p BluRay].mkv", "", 1, 3],
-            ["Revolution.2012.S02E14.720p.HDTV.X264-DIMENSION [PublicHD].mkv", "Revolution 2012", 2, 14],
-            ["Castle.2009.S06E03.720p.HDTV.X264-DIMENSION [PublicHD].mkv", "Castle 2009", 6, 3],
-            ["19-2.2014.S02E01.720p.HDTV.x264-CROOKS", "19-2 2014", 2, 1],
+            ["Revolution.2012.S02E14.720p.HDTV.X264-DIMENSION [PublicHD].mkv", "Revolution", 2, 14],
+            ["Castle.2009.S06E03.720p.HDTV.X264-DIMENSION [PublicHD].mkv", "Castle", 6, 3],
+            ["19-2.2014.S02E01.720p.HDTV.x264-CROOKS", "19-2", 2, 1],
             ["Community - S01E09 - Debate 109", "Community", 1, 9],
             ["Entourage - S02E02 - My Maserati Does 185", "Entourage", 2, 2],
             ["6x13 - The Family Guy 100th Episode Special", "", 6, 13],
@@ -109,7 +109,7 @@ class TVMatch(TestCase):
             # TODO - handle absolute episode parsing (i.e 103)
             #["quantico.103.hdtv-lol[ettv].mp4", "quantico", 1, 3],
             ["Fargo - 01x02 - The Rooster Prince - [itz_theo]", "Fargo", 1, 2],
-            ["Castle (2009) - [06x16] - Room 147.mp4", "Castle (2009)", 6, 16],
+            ["Castle (2009) - [06x16] - Room 147.mp4", "Castle", 6, 16],
             ["grp-zoos01e11-1080p", "grp-zoo", 1, 11],
             ["grp-zoo-s01e11-1080p", "grp-zoo", 1, 11],
             ["Jeopardy!.S2016E14.2016-01-20.avi", "Jeopardy!", 2016, 14],
@@ -118,7 +118,7 @@ class TVMatch(TestCase):
             ["Judge Judy 2016 02 25 S20E143", "Judge Judy", 20, 143],
             ["Red Dwarf - S02 - E06 - Parallel Universe", "Red Dwarf", 2, 6],
             ["O.J.Simpson.Made.in.America.Part.Two.720p.HDTV.x264-2HD", "O J Simpson Made in America", 1, 2],
-            ["The.100000.Dollar.Pyramid.2016.S01E05.720p.HDTV.x264-W4F", "The 100000 Dollar Pyramid 2016", 1, 5],
+            ["The.100000.Dollar.Pyramid.2016.S01E05.720p.HDTV.x264-W4F", "The 100000 Dollar Pyramid", 1, 5],
             ["Class S01E02 (22 October 2016) HDTV 720p [Webrip]", "Class", 1, 2],
             #["this.is.not.happening.2015.0308-yestv", "this is not happening 2015", 3, 8],
             ["Jeopardy - S2016E231", "Jeopardy", 2016, 231],
